Fix some 32-bit builds after r212699
* platform/mediastream/mac/RealtimeIncomingVideoSource.cpp:
(WebCore::RealtimeIncomingVideoSource::OnFrame):
Explicitly cast OSStatus to int for logging.



Canonical link: https://commits.webkit.org/185639@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212701 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/mediastream/mac/RealtimeIncomingVideoSource.cpp
@@ -110,14 +110,14 @@ void RealtimeIncomingVideoSource::OnFrame(const webrtc::VideoFrame& frame)
     CMVideoFormatDescriptionRef formatDescription;
     OSStatus ostatus = CMVideoFormatDescriptionCreateForImageBuffer(kCFAllocatorDefault, (CVImageBufferRef)pixelBuffer, &formatDescription);
     if (ostatus != noErr) {
-        LOG_ERROR("Failed to initialize CMVideoFormatDescription: %d", ostatus);
+        LOG_ERROR("Failed to initialize CMVideoFormatDescription: %d", static_cast<int>(ostatus));
         return;
     }
 
     CMSampleBufferRef sampleBuffer;
     ostatus = CMSampleBufferCreateReadyWithImageBuffer(kCFAllocatorDefault, (CVImageBufferRef)pixelBuffer, formatDescription, &timingInfo, &sampleBuffer);
     if (ostatus != noErr) {
-        LOG_ERROR("Failed to create the sample buffer: %d", ostatus);
+        LOG_ERROR("Failed to create the sample buffer: %d", static_cast<int>(ostatus));
         return;
     }
     CFRelease(formatDescription);
